# api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: CLM-Core
  version: 1.0.0
  description: API endpoints the clm-core module offers
servers:
  - url: '{scheme}://{hostname}:{port}{path}'
    description: The production API server
    variables:
      hostname:
        default: localhost
      port:
        default: '5008'
      path:
        default: ''
      scheme:
        default: http
security:
  - bearerAuth: []
paths:
...
  '/core/mgmt/consumers/{encryptedId}/confirm':
    get:
      tags:
        - pblc-redirect
      description: Double-opt in for API-Token when created.
      summary: Confiming API-Token
      parameters:
        - in: path
          name: encryptedId
          required: true
          description: The encrypted id of the API-Token
          schema:
            type: string
      responses:
        '200':
          description: Successfully activated API-Token
  /core/mgmt/consumers:
    get:
      tags:
        - mgmt-consumers
        - mgmt
      description: Super-Admin can get all API-Tokens
      summary: 'Get all API-Tokens [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: All API-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/consumer'
    post:
      tags:
        - mgmt-consumers
        - mgmt
      description: Super-Admin can get create API-Tokens
      summary: 'Create an API-Token [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully created API-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/consumer'
  '/core/mgmt/consumers/{id}':
    put:
      tags:
        - mgmt-consumers
        - mgmt
      description: Super-Admin can get update API-Tokens
      summary: 'Update an API-Token [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
        - in: path
          name: id
          required: true
          description: The id of the API-Token
          schema:
            type: string
      responses:
        '200':
          description: Successfully updated API-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/consumer'
...
components:
  schemas:
...
    consumer:
      properties:
        displayName:
          type: string
          default: API-Token
        active:
          type: boolean
          default: false
        userId:
          type: boolean
          description: The user who associates with that API-Token
          default: consumer@localhost.org
        domain:
          type: string
          default: STAKEHOLDER-DOMAIN
        paths:
          type: object
          properties:
            scope:
              type: array
              items:
                type: string
                enum:
                  - GET
                  - PUT
                  - POST
                  - DELETE
                  - PATCH
              description: The HTTP methods which can be applied to the route
            route:
              type: string
              default: /some/route/the/token/can/access
...
  parameters:
    accessToken:
      name: x-access-token
      in: header
      description: The access token
      required: true
      example: exampleAccessToken
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshAuth:
      type: apiKey
      in: header
      name: x-refresh-token
...